#10: refactors all components to use lombok annotations, refactors the discord bot token injection to be handled by spring boot
