+++ src/main/java/com/github/daci1/discord_bot/commands/audio/player/commands/NowPlayingCommand.java
@@ -2,30 +2,26 @@ package com.github.daci1.discord_bot.commands.audio.player.commands;
 
 import com.github.daci1.discord_bot.services.MembersStateService;
 import com.github.daci1.discord_bot.services.PlayerManagerService;
-import com.github.daci1.discord_bot.DiscordBotService;
+import com.github.daci1.discord_bot.services.DiscordBotService;
 import com.github.daci1.discord_bot.commands.ISlashCommand;
 import com.github.daci1.discord_bot.commands.SlashCommand;
 
 import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+
 @Component
+@RequiredArgsConstructor
 public class NowPlayingCommand extends ListenerAdapter implements ISlashCommand {
 
-    @Autowired
-    private DiscordBotService discordBotService;
-
-    @Autowired
-    private MembersStateService membersStateService;
-
-    @Autowired
-    private PlayerManagerService playerManagerService;
-
+    private final DiscordBotService discordBotService;
+    private final MembersStateService membersStateService;
+    private final PlayerManagerService playerManagerService;
 
     @Override
     public void onSlashCommandInteraction(SlashCommandInteractionEvent event) {
