+++ src/main/java/com/github/daci1/discord_bot/commands/audio/player/commands/QueueCommand.java
@@ -1,24 +1,22 @@
 package com.github.daci1.discord_bot.commands.audio.player.commands;
 
-import com.github.daci1.discord_bot.DiscordBotService;
+import com.github.daci1.discord_bot.services.DiscordBotService;
 import com.github.daci1.discord_bot.commands.ISlashCommand;
 import com.github.daci1.discord_bot.commands.SlashCommand;
 
 import com.github.daci1.discord_bot.services.PlayerManagerService;
 import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
+@RequiredArgsConstructor
 public class QueueCommand extends ListenerAdapter implements ISlashCommand {
 
-    @Autowired
-    private PlayerManagerService playerManagerService;
-
-    @Autowired
-    private DiscordBotService discordBotService;
+    private final PlayerManagerService playerManagerService;
+    private final DiscordBotService discordBotService;
 
     @Override
     public void onSlashCommandInteraction(SlashCommandInteractionEvent event) {
