+++ src/main/java/com/github/daci1/discord_bot/commands/audio/player/commands/RepeatCommand.java
@@ -3,28 +3,24 @@ package com.github.daci1.discord_bot.commands.audio.player.commands;
 import com.github.daci1.discord_bot.commands.CommandUtils;
 import com.github.daci1.discord_bot.services.MembersStateService;
 import com.github.daci1.discord_bot.services.PlayerManagerService;
-import com.github.daci1.discord_bot.DiscordBotService;
+import com.github.daci1.discord_bot.services.DiscordBotService;
 import com.github.daci1.discord_bot.commands.ISlashCommand;
 import com.github.daci1.discord_bot.commands.SlashCommand;
 import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
+@RequiredArgsConstructor
 public class RepeatCommand extends ListenerAdapter implements ISlashCommand {
 
-    @Autowired
-    private DiscordBotService discordBotService;
-
-    @Autowired
-    private PlayerManagerService playerManagerService;
-
-    @Autowired
-    private MembersStateService membersStateService;
+    private final DiscordBotService discordBotService;
+    private final PlayerManagerService playerManagerService;
+    private final MembersStateService membersStateService;
 
     @Override
     public void onSlashCommandInteraction(SlashCommandInteractionEvent event) {
